requirements.txt
flask==3.0.0
flask-cors==4.0.0
resend==0.7.2
requests==2.31.0
gunicorn==21.2.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
